Implement extension for create the karaoke hit-object specific json setting.
