+++ osu.Game.Rulesets.Karaoke/Beatmaps/Formats/KaraokeJsonBeatmapEncoder.cs
@@ -5,8 +5,7 @@ using System;
 using System.Linq;
 using Newtonsoft.Json;
 using osu.Game.Beatmaps;
-using osu.Game.IO.Serialization;
-using osu.Game.Rulesets.Karaoke.IO.Serialization.Converters;
+using osu.Game.Rulesets.Karaoke.IO.Serialization;
 using osu.Game.Rulesets.Karaoke.Objects;
 
 namespace osu.Game.Rulesets.Karaoke.Beatmaps.Formats
@@ -15,15 +14,7 @@ namespace osu.Game.Rulesets.Karaoke.Beatmaps.Formats
     {
         public string Encode(Beatmap output)
         {
-            var globalSetting = JsonSerializableExtensions.CreateGlobalSettings();
-            globalSetting.Converters.Add(new CultureInfoConverter());
-            globalSetting.Converters.Add(new RomajiTagConverter());
-            globalSetting.Converters.Add(new RomajiTagsConverter());
-            globalSetting.Converters.Add(new RubyTagConverter());
-            globalSetting.Converters.Add(new RubyTagsConverter());
-            globalSetting.Converters.Add(new TimeTagConverter());
-            globalSetting.Converters.Add(new TimeTagsConverter());
-            globalSetting.Converters.Add(new ToneConverter());
+            var globalSetting = KaraokeJsonSerializableExtensions.CreateGlobalSettings();
 
             // should check that every parent in the note contains in the hit objects list.
             var lyrics = output.HitObjects.OfType<Lyric>();
